add button block deprecation
