+++ src/blocks/button/deprecated/deprecated.js
@@ -1,0 +1,7 @@
+// Import earlier versions of the block.
+import { config as v1 } from './v1/v1.js';
+import { config as v2 } from './v2/v2.js';
+
+const deprecated = [ v2, v1 ];
+
+export default deprecated;
